DESS: calculate energy values for chargers, pv-meter and grid-meter
