+++ io.openems.edge.fenecon.dess/src/io/openems/edge/fenecon/dess/pvmeter/FeneconDessPvMeter.java
@@ -1,52 +1,15 @@
 package io.openems.edge.fenecon.dess.pvmeter;
 
-import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.ConfigurationPolicy;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
-import org.osgi.service.metatype.annotations.Designate;
+import org.osgi.service.event.EventHandler;
 
-import io.openems.common.exceptions.OpenemsException;
-import io.openems.edge.bridge.modbus.api.AbstractOpenemsModbusComponent;
-import io.openems.edge.bridge.modbus.api.BridgeModbus;
-import io.openems.edge.bridge.modbus.api.ElementToChannelConverter;
-import io.openems.edge.bridge.modbus.api.ModbusProtocol;
-import io.openems.edge.bridge.modbus.api.element.UnsignedWordElement;
-import io.openems.edge.bridge.modbus.api.task.FC3ReadRegistersTask;
 import io.openems.edge.common.channel.Doc;
 import io.openems.edge.common.component.OpenemsComponent;
-import io.openems.edge.common.taskmanager.Priority;
-import io.openems.edge.fenecon.dess.FeneconDessConstants;
 import io.openems.edge.meter.api.AsymmetricMeter;
-import io.openems.edge.meter.api.MeterType;
 import io.openems.edge.meter.api.SymmetricMeter;
+import io.openems.edge.timedata.api.TimedataProvider;
 
-@Designate(ocd = Config.class, factory = true)
-@Component(//
-		name = "Fenecon.Dess.PvMeter", //
-		immediate = true, //
-		configurationPolicy = ConfigurationPolicy.REQUIRE, //
-		property = { //
-				"type=PRODUCTION" //
-		})
-public class FeneconDessPvMeter extends AbstractOpenemsModbusComponent
-		implements AsymmetricMeter, SymmetricMeter, OpenemsComponent {
-
-	@Reference
-	protected ConfigurationAdmin cm;
-
-	@Reference(policy = ReferencePolicy.STATIC, policyOption = ReferencePolicyOption.GREEDY, cardinality = ReferenceCardinality.MANDATORY)
-	protected void setModbus(BridgeModbus modbus) {
-		super.setModbus(modbus);
-	}
-
-	private String modbusBridgeId;
+public interface FeneconDessPvMeter
+		extends AsymmetricMeter, SymmetricMeter, OpenemsComponent, TimedataProvider, EventHandler {
 
 	public enum ChannelId implements io.openems.edge.common.channel.ChannelId {
 		;
@@ -61,73 +24,5 @@ public class FeneconDessPvMeter extends AbstractOpenemsModbusComponent
 			return this.doc;
 		}
 	}
-
-	public FeneconDessPvMeter() {
-		super(//
-				OpenemsComponent.ChannelId.values(), //
-				SymmetricMeter.ChannelId.values(), //
-				AsymmetricMeter.ChannelId.values(), //
-				ChannelId.values() //
-		);
-
-		// automatically calculate Active/ReactivePower from L1/L2/L3
-		AsymmetricMeter.initializePowerSumChannels(this);
-	}
-
-	@Activate
-	void activate(ComponentContext context, Config config) throws OpenemsException {
-		if (super.activate(context, config.id(), config.alias(), config.enabled(), FeneconDessConstants.UNIT_ID,
-				this.cm, "Modbus", config.modbus_id())) {
-			return;
-		}
-		this.modbusBridgeId = config.modbus_id();
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		super.deactivate();
-	}
-
-	public String getModbusBridgeId() {
-		return this.modbusBridgeId;
-	}
-
-	@Override
-	protected ModbusProtocol defineModbusProtocol() throws OpenemsException {
-		return new ModbusProtocol(this, //
-				new FC3ReadRegistersTask(11144, Priority.HIGH, //
-						m(AsymmetricMeter.ChannelId.ACTIVE_POWER_L1, new UnsignedWordElement(11144), DELTA_10000)), //
-				new FC3ReadRegistersTask(11174, Priority.HIGH, //
-						m(AsymmetricMeter.ChannelId.ACTIVE_POWER_L2, new UnsignedWordElement(11174), DELTA_10000)), //
-				new FC3ReadRegistersTask(11204, Priority.HIGH, //
-						m(AsymmetricMeter.ChannelId.ACTIVE_POWER_L3, new UnsignedWordElement(11204), DELTA_10000)) //
-		);
-	}
-
-	@Override
-	public MeterType getMeterType() {
-		return MeterType.PRODUCTION;
-	}
-
-	@Override
-	public String debugLog() {
-		return "P:" + this.getActivePower().asString();
-	}
-
-	private static final ElementToChannelConverter DELTA_10000 = new ElementToChannelConverter(//
-			// element -> channel
-			value -> {
-				if (value == null) {
-					return null;
-				}
-				int intValue = (Integer) value;
-				if (intValue == 0) {
-					return 0; // ignore '0'
-				}
-				return (intValue - 10_000); // apply delta of 10_000
-			}, //
-
-			// channel -> element
-			value -> value);
 
 }
