+++ io.openems.edge.fenecon.dess/test/io/openems/edge/fenecon/dess/gridmeter/FeneconDessGridMeterTest.java
@@ -13,7 +13,7 @@ public class FeneconDessGridMeterTest {
 
 	@Test
 	public void test() throws Exception {
-		new ComponentTest(new FeneconDessGridMeter()) //
+		new ComponentTest(new FeneconDessGridMeterImpl()) //
 				.addReference("cm", new DummyConfigurationAdmin()) //
 				.addReference("setModbus", new DummyModbusBridge(MODBUS_ID)) //
 				.activate(MyConfig.create() //
